🐛 fix(embed): View を別モジュール化して Fast Refresh 破損を解消
plugin.tsx が factory（非コンポーネント）と React コンポーネント（EmbedView/
DefaultEmbedChrome）を同一モジュールで export していたため vite-plugin-react の
Fast Refresh が無効化され（"createEmbedShapePlugin export is incompatible"）、
HMR のたびに旧 render クロージャが残り rt undefined でクラッシュ→「エラーで
表示されない」の原因だった。

- embed-view.tsx: コンポーネントのみ（EmbedView / DefaultEmbedChrome / 型）に分離。
- embed-events.ts: window イベントの ACTION 定数/型/emit を分離（非コンポーネント）。
- plugin.tsx: factory + 図形登録のみ（コンポーネント export なし）。
- EmbedView に rt 未定義ガードを追加（万一の stale 描画でも throw しない）。
